Two bug fixes in URL handling:
(1) Setting a query string argument to null or undefined will no longer that
argument to appear as "v=null" or "v=undefined" in the URL.

(2) Incoming query string values equal to the empty string will be
converted to nulls. E.g. if the query string is "a=&b=null", the
resulting arguments dictionary will now be {a: null, b: "null"}.